[MC-15] [api] refactor: change import path name and update adapter logic to align with the refactored user domain logic.
- Rename utils directory to commons
- Change directory name from auth to user in the user domain area
- Change from AuthService to UserMetadataRepository code
- Change mockdata class type from MemberEntity to UserMetadataEntity
- Add Authorization header in test code for JWT token pass
